+++ listen-kit/src/dexscreener/mod.rs
@@ -1,69 +1,9 @@
 pub mod tools;
+pub mod types;
 
 use anyhow::Result;
 use reqwest::Client;
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct DexScreenerResponse {
-    #[serde(rename = "schemaVersion")]
-    pub schema_version: String,
-    pub pairs: Vec<PairInfo>,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct PairInfo {
-    #[serde(rename = "chainId")]
-    pub chain_id: String,
-    #[serde(rename = "dexId")]
-    pub dex_id: String,
-    pub url: String,
-    #[serde(rename = "pairAddress")]
-    pub pair_address: String,
-    pub labels: Option<Vec<String>>,
-    #[serde(rename = "baseToken")]
-    pub base_token: Token,
-    #[serde(rename = "quoteToken")]
-    pub quote_token: Token,
-    #[serde(rename = "priceNative")]
-    pub price_native: String,
-    #[serde(rename = "priceUsd")]
-    pub price_usd: Option<String>,
-    pub liquidity: Option<Liquidity>,
-    pub volume: Option<Volume>,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct Liquidity {
-    pub usd: Option<f64>,
-    // pub base: Option<f64>,
-    // pub quote: Option<f64>,
-    // not relevant for us, save tokens
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone, Default)]
-pub struct Volume {
-    #[serde(default)]
-    pub h24: Option<f64>,
-    // #[serde(default)]
-    // pub h6: Option<f64>,
-    // #[serde(default)]
-    // pub h1: Option<f64>,
-    // #[serde(default)]
-    // pub m5: Option<f64>,
-    // not relevant for us, save tokens
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct Token {
-    pub address: String,
-    pub name: String,
-    pub symbol: String,
-}
-
-pub struct TickerResponse {
-    pub mint: String,
-}
+use types::DexScreenerResponse;
 
 pub async fn search_ticker(ticker: String) -> Result<DexScreenerResponse> {
     let client = Client::new();
@@ -84,8 +24,28 @@ pub async fn search_ticker(ticker: String) -> Result<DexScreenerResponse> {
 
     let mut dex_response: DexScreenerResponse = serde_json::from_value(data)?;
 
-    // trim up to 6
-    dex_response.pairs.truncate(6);
+    // sort by liquidity (highest first)
+    dex_response.pairs.sort_by(|a, b| {
+        let a_liq = if a.liquidity.is_some() {
+            a.liquidity.as_ref().unwrap().usd.unwrap_or(0.0)
+        } else {
+            0.0
+        };
+
+        let b_liq = if b.liquidity.is_some() {
+            b.liquidity.as_ref().unwrap().usd.unwrap_or(0.0)
+        } else {
+            0.0
+        };
+
+        // Compare in reverse order (b compared to a) for descending sort
+        b_liq
+            .partial_cmp(&a_liq)
+            .unwrap_or(std::cmp::Ordering::Equal)
+    });
+
+    // take top 6
+    dex_response.pairs.truncate(5);
 
     Ok(dex_response)
 }
@@ -105,7 +65,7 @@ mod tests {
     async fn test_search_ticker() {
         let response = search_ticker("BONK".to_string()).await.unwrap();
         assert_eq!(response.schema_version, "1.0.0");
-        println!("{:?}", response);
+        println!("{:#?}", response);
     }
 
     #[tokio::test]
